ml_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
xgboost
catboost
scikit-learn
pandas
numpy
shap
joblib
python-multipart

google-generativeai==0.3.2
pgvector==0.2.4
